rewrite safe stream without init
Remove Init and InitAck methods. Make a bigger Send method with
begin_offset and DataMessage contains delta offset with repeat count and
reset flag

+++ aether/stream_api/safe_stream/safe_stream_config.h
@@ -26,10 +26,10 @@ namespace ae {
 struct SafeStreamConfig {
   SSRingIndex::type buffer_capacity;  //< sending buffer capacity
   SSRingIndex::type window_size;      //< size of sending window
-  SSRingIndex::type max_packet_size;    //< max size of sending data
-  std::uint16_t max_repeat_count;  //< max repeat count for sending packet
-  Duration wait_confirm_timeout;   //< Timeout for waiting confirmation
-  Duration send_confirm_timeout;   //< max time to wait before send confirmation
+  SSRingIndex::type max_packet_size;  //< max size of sending data
+  std::uint8_t max_repeat_count;      //< max repeat count for sending packet
+  Duration wait_confirm_timeout;      //< Timeout for waiting confirmation
+  Duration send_confirm_timeout;  //< max time to wait before send confirmation
   Duration send_repeat_timeout;  //< max time to wait before send repeat request
 };
